+++ src/template/js/src/client/App.js
@@ -1,11 +1,29 @@
 import * as React from "react";
 import logo from "./logo.svg";
+import BeerDisplayer from "./components/BeerDisplayer";
+import { initializeStarterBeers, getBeers } from "./api/beer";
 
 const App = () => {
+  const [beers, setBeers] = React.useState([]);
+
+  React.useEffect(() => {
+    const initializeData = async () => {
+      try {
+        let currentBeers = await getBeers();
+        if (currentBeers.length === 0) {
+          await initializeStarterBeers();
+          currentBeers = await getBeers();
+        }
+        setBeers(currentBeers);
+      } catch (error) {}
+    };
+    initializeData();
+  }, []);
+
   return (
     <div id="app">
       <img alt="Logo" id="logo" src={logo} />
-      <h1>create-mern-application</h1>
+      <h1>___APP NAME___</h1>
       <p>
         Edit <code>src/client/App.tsx</code> and save to reload.
       </p>
@@ -16,6 +34,7 @@ const App = () => {
       >
         View Documentation
       </a>
+      <BeerDisplayer beers={beers} />
     </div>
   );
 };