+++ src/template/ts/src/client/types.ts
@@ -1,0 +1,9 @@
+export interface Beer {
+  __v: string;
+  _id: string;
+  abv: number;
+  brewer: string;
+  description: string;
+  name: string;
+  type: string;
+}