+++ CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 2020-06-05
+
+### 🍻 Beers in the DB
+
+### Added
+
+- Beers Database [#10](https://github.com/alexlee-dev/create-mern-application/issues/10)
+
+### Changed
+
+- Add application name to `App.js` / `App.tsx`
+
+### Removed
+
+### Fixed
+
+- No assets being served by server
+
 ## [0.4.0] - 2020-06-04
 
 ### 🧪 Testing